requirements.txt
grpcio==1.66.2
grpcio-tools==1.66.2
protobuf==5.28.2
setuptools==75.1.0
